* updated default conversion to byte array * updating some model types

+++ src/Foundation/MSSDK/code/Decision/Models/Personalizer/Ranking.cs
@@ -8,6 +8,6 @@ namespace SitecoreCognitiveServices.Foundation.MSSDK.Decision.Models.Personalize
     public class Ranking
     {
         public string id { get; set; }
-        public float probability { get; set; }
+        public double probability { get; set; }
     }
 }
